Rename namespace after merge from main
